refactor: split CharacterPreview test file and implement missing functionality using TDD
- Split 569-line CharacterPreview.test.tsx into focused test files:
  - CharacterPreview.rendering.test.tsx
  - CharacterPreview.basicInfo.test.tsx
  - CharacterPreview.classes.test.tsx
  - CharacterPreview.abilities.test.tsx
  - CharacterPreview.combat.test.tsx
  - CharacterPreview.validation.test.tsx
- Created reusable test helpers in CharacterPreview.helpers.ts
- Implemented missing validation completion status functionality in CharacterPreview component
- Fixed race name formatting to properly handle hyphenated races (Half-Elf, Half-Orc)
- Added comprehensive validation status display with ✓/⚠ indicators
- All test files now under 450-line coding standard limit
- Follows TDD principles by implementing functionality to pass tests

Relates to #181
