+++ .codacy/logs/codacy-cli.log
@@ -105,3 +105,12 @@
 2025-07-07T16:21:16-07:00 [DEBUG] (codacy-cli-v2/tools/pmdRunner.go:160) Running PMD command command=/home/doug/.cache/codacy/tools/pmd@6.55.0/pmd-bin-6.55.0/bin/run.sh pmd -R .codacy/tools-configs/ruleset.xml -d /home/doug/third-dev/dnd-tracker-next-js -f sarif -r /tmp/codacy-analysis-251352628/pmd.sarif
 2025-07-07T16:22:05-07:00 [INFO] (codacy-cli-v2/cli-v2.go:27) Starting Codacy CLI version=1.0.0-main.339.sha.dcd97cf (dcd97cf) built at 2025-06-16T20:10:29Z
 2025-07-07T16:22:05-07:00 [INFO] (codacy-cli-v2/cmd/root.go:34) Executing CLI command full_command=[codacy-cli analyze --tool eslint --format sarif --output codacy-eslint-only.sarif] args=[] command=analyze
+2025-07-07T19:01:45-07:00 [INFO] (codacy-cli-v2/cli-v2.go:27) Starting Codacy CLI version=1.0.0-main.339.sha.dcd97cf (dcd97cf) built at 2025-06-16T20:10:29Z
+2025-07-07T19:01:45-07:00 [INFO] (codacy-cli-v2/cmd/root.go:34) Executing CLI command command=analyze full_command=[/home/doug/.cache/codacy/codacy-cli-v2/1.0.0-main.339.sha.dcd97cf/codacy-cli-v2 analyze --format sarif --output codacy-current-pr-analysis.sarif] args=[]
+2025-07-07T19:01:46-07:00 [DEBUG] (codacy-cli-v2/tools/revive/reviveRunner.go:53) Running Revive command command=/home/doug/.cache/codacy/tools/revive@1.7.0/revive -config .codacy/tools-configs/revive.toml -formatter sarif ./...
+2025-07-07T19:03:23-07:00 [DEBUG] (codacy-cli-v2/tools/pmdRunner.go:32) PMD binary path pmdBinary=/home/doug/.cache/codacy/tools/pmd@6.55.0/pmd-bin-6.55.0/bin/run.sh
+2025-07-07T19:03:23-07:00 [DEBUG] (codacy-cli-v2/tools/pmdRunner.go:98) Setting up Java environment javaHome=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7
+2025-07-07T19:03:23-07:00 [DEBUG] (codacy-cli-v2/tools/pmdRunner.go:144) Updated environment variables javaHome=JAVA_HOME=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7 path=PATH=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7/bin:/home/doug/bin:/home/doug/.npm-global/bin:/home/doug/bin:/home/doug/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin:/snap/bin binDir=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7/bin javaBinary=/home/doug/.cache/codacy/runtimes/jdk-17.0.10+7/bin/java
+2025-07-07T19:03:23-07:00 [DEBUG] (codacy-cli-v2/tools/pmdRunner.go:160) Running PMD command command=/home/doug/.cache/codacy/tools/pmd@6.55.0/pmd-bin-6.55.0/bin/run.sh pmd -R .codacy/tools-configs/ruleset.xml -d /home/doug/third-dev/dnd-tracker-next-js -f sarif -r /tmp/codacy-analysis-3649038988/pmd.sarif
+2025-07-07T19:04:32-07:00 [INFO] (codacy-cli-v2/cli-v2.go:27) Starting Codacy CLI version=1.0.0-main.339.sha.dcd97cf (dcd97cf) built at 2025-06-16T20:10:29Z
+2025-07-07T19:04:32-07:00 [INFO] (codacy-cli-v2/cmd/root.go:34) Executing CLI command command=analyze full_command=[/home/doug/.cache/codacy/codacy-cli-v2/1.0.0-main.339.sha.dcd97cf/codacy-cli-v2 analyze --tool eslint --format sarif --output codacy-eslint-analysis.sarif src/components/forms/character/] args=[src/components/forms/character/]
