+++ codacy-eslint-analysis.sarif
@@ -1,0 +1,5 @@
+{
+  "$schema": "https://raw.githubusercontent.com/oasis-tcs/sarif-spec/master/Schemata/sarif-schema-2.1.0.json",
+  "runs": [],
+  "version": "2.1.0"
+}
